fix webview resize, extract 'MiniBrowserMouseClickTracker' as 'ApplicationShellMouseTracker' in @theia/core
