+++ packages/core/src/browser/frontend-application-module.ts
@@ -73,6 +73,7 @@ import { ContextKeyService } from './context-key-service';
 import { ResourceContextKey } from './resource-context-key';
 import { KeyboardLayoutService } from './keyboard/keyboard-layout-service';
 import { MimeService } from './mime-service';
+import { ApplicationShellMouseTracker } from './shell/application-shell-mouse-tracker';
 
 export const frontendApplicationModule = new ContainerModule((bind, unbind, isBound, rebind) => {
     const themeService = ThemeService.get();
@@ -235,6 +236,9 @@ export const frontendApplicationModule = new ContainerModule((bind, unbind, isBo
     bindCorePreferences(bind);
 
     bind(MimeService).toSelf().inSingletonScope();
+
+    bind(ApplicationShellMouseTracker).toSelf().inSingletonScope();
+    bind(FrontendApplicationContribution).toService(ApplicationShellMouseTracker);
 });
 
 export function bindMessageService(bind: interfaces.Bind): interfaces.BindingWhenOnSyntax<MessageService> {
